% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataviz.R
\name{sankey_ly}
\alias{sankey_ly}
\title{Sankey Plot with Plotly}
\usage{
sankey_ly(x, cols, num_col)
}
\arguments{
\item{x}{A data.frame input}

\item{cols}{A vector of columns names}

\item{num_col}{The numeric column name}
}
\description{
Sankey Plot with Plotly
}
\details{
A customized function for plotting sankey plots with Plotly
}
