% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataviz.R
\name{sankey_ly}
\alias{sankey_ly}
\title{Sankey Plot with Plotly}
\usage{
sankey_ly(x, cat_cols, num_col, title = NULL)
}
\arguments{
\item{x}{A data.frame input, must have at least two categorical columns and one numeric column}

\item{cat_cols}{A  vector of at least two categorical columns names}

\item{num_col}{A single numeric column name}

\item{title}{Optional, string to pass to plotly layout title function}
}
\description{
Sankey Plot with Plotly
}
\details{
A customized function for data transformation and plotting sankey plot with Plotly
}
